+++ lib/screens/browser.dart
@@ -224,6 +224,7 @@ class _BrowserPageState extends ConsumerState<BrowserPage> {
                       builder: (context) => MapChooserPage(
                         creating: true,
                         location: location,
+                        closer: editorMode == EditorMode.micromapping,
                       ),
                     ),
                   );